add --kindle option to optimize for display on Kindle
Add a command-line option to optimize the HTML and CSS for display on
Kindle eInk devices. Currently one optimization is made: a CSS style
to display <p> tags inside <li> tags as inline-block rather than
block. This fixes a formatting issue on the Kindle under which the
device displays an extra newline after between the bullet point and
the text of the list item.