inline looks the same on Kindle and better in ebook-viewer
Change the <li> <p> inline-block Kindle optimization to inline. This
has the same effect as inline-block on the Kindle (i.e. it fixes the
extra-newline issue), but it looks better in ebook-viewer and possibly
on other devices (in ebook-viewer, inline-block causes the bullet
point to align with the LAST line in the paragraph, rather than the
first.